Refactor auth forms to separate components
