+++ src/components/Auth/separator.tsx
@@ -1,0 +1,11 @@
+const Separator = () => {
+  return (
+    <div className="m-4 flex items-center">
+      <div className="flex-1 border-t-2 border-dashed border-primary-600"></div>
+      <div className="mx-4 text-primary-600">or</div>
+      <div className="flex-1 border-t-2 border-dashed border-primary-600"></div>
+    </div>
+  );
+};
+
+export default Separator;
