+++ src/components/Button/sign-in-oauth.tsx
@@ -1,9 +1,16 @@
+import { ReactNode } from "react";
 import { useSignIn } from "@clerk/nextjs";
 import { OAuthStrategy } from "@clerk/nextjs/dist/types/server";
 
 import Button from "@/components/Button";
 
-const SignInOAuthButton = () => {
+interface Props {
+  strategy: OAuthStrategy;
+  label: string;
+  icon: ReactNode;
+}
+
+const SignInOAuthButton = ({ strategy, label, icon }: Props) => {
   const { signIn } = useSignIn();
 
   const signInWith = (strategy: OAuthStrategy) => {
@@ -14,12 +21,13 @@ const SignInOAuthButton = () => {
     });
   };
 
-  // Render a button for each supported OAuth provider
-  // you want to add to your app
   return (
-    <div>
-      <Button onClick={() => signInWith("oauth_google")}>
-        Sign in with Google
+    <div className="flex justify-center">
+      <Button onClick={() => signInWith(strategy)} width="w-full">
+        <div className="relative flex items-center justify-center space-x-4">
+          <div className="w-5">{icon}</div>
+          <span className="">{label}</span>
+        </div>
       </Button>
     </div>
   );
