+++ src/middleware.ts
@@ -2,7 +2,7 @@ import { authMiddleware } from "@clerk/nextjs";
 
 export default authMiddleware({
   // url routes that don't require login
-  publicRoutes: ["/login", "/sso-callback"]
+  publicRoutes: ["/login", "/sso-callback", "/forgot-password"]
 });
 
 export const config = {
